+++ Source/Scene/SkyBox.js
@@ -20,7 +20,7 @@ define([
         destroyObject,
         DeveloperError,
         Matrix4,
-        MeshFilters,
+        GeometryFilters,
         PrimitiveType,
         loadCubeMap,
         BufferUsage,
@@ -151,7 +151,7 @@ define([
             var mesh = BoxTessellator.compute({
                 dimensions : new Cartesian3(2.0, 2.0, 2.0)
             });
-            var attributeIndices = MeshFilters.createAttributeIndices(mesh);
+            var attributeIndices = GeometryFilters.createAttributeIndices(mesh);
 
             command.primitiveType = PrimitiveType.TRIANGLES;
             command.modelMatrix = Matrix4.IDENTITY.clone();