+++ Specs/Core/GeometryFiltersSpec.js
@@ -1,6 +1,6 @@
 /*global defineSuite*/
 defineSuite([
-         'Core/MeshFilters',
+         'Core/GeometryFilters',
          'Core/PrimitiveType',
          'Core/ComponentDatatype',
          'Core/EllipsoidGeometry',
@@ -13,7 +13,7 @@ defineSuite([
          'Core/GeometryAttribute',
          'Core/GeometryIndices'
      ], function(
-         MeshFilters,
+         GeometryFilters,
          PrimitiveType,
          ComponentDatatype,
          EllipsoidGeometry,
@@ -29,7 +29,7 @@ defineSuite([
     /*global jasmine,describe,xdescribe,it,xit,expect,beforeEach,afterEach,beforeAll,afterAll,spyOn,runs,waits,waitsFor*/
 
     it('converts triangles to wireframe in place', function() {
-        var mesh = MeshFilters.toWireframeInPlace({
+        var mesh = GeometryFilters.toWireframe({
             indexLists : [{
                 primitiveType : PrimitiveType.TRIANGLES,
                 values : [0, 1, 2, 3, 4, 5]
@@ -57,7 +57,7 @@ defineSuite([
     });
 
     it('converts a triangle fan to wireframe in place', function() {
-        var mesh = MeshFilters.toWireframeInPlace({
+        var mesh = GeometryFilters.toWireframe({
             indexLists : [{
                 primitiveType : PrimitiveType.TRIANGLE_FAN,
                 values : [0, 1, 2, 3]
@@ -85,7 +85,7 @@ defineSuite([
     });
 
     it('converts a triangle strip to wireframe in place', function() {
-        var mesh = MeshFilters.toWireframeInPlace({
+        var mesh = GeometryFilters.toWireframe({
             indexLists : [{
                 primitiveType : PrimitiveType.TRIANGLE_STRIP,
                 values : [0, 1, 2, 3]
@@ -121,7 +121,7 @@ defineSuite([
             }
         };
 
-        var indices = MeshFilters.createAttributeIndices(mesh);
+        var indices = GeometryFilters.createAttributeIndices(mesh);
 
         var validIndices = [0, 1, 2];
         expect(validIndices).toContain(indices.position);
@@ -138,7 +138,7 @@ defineSuite([
             colors : 2
         };
 
-        var mappedIndices = MeshFilters.mapAttributeIndices(indices, {
+        var mappedIndices = GeometryFilters.mapAttributeIndices(indices, {
             positions : 'position',
             normals : 'normal',
             colors : 'color'
@@ -165,7 +165,7 @@ defineSuite([
                 componentsPerAttribute : 3,
                 values : [0, 1, 2, 3, 4, 5]
             };
-            mesh = MeshFilters.reorderForPreVertexCache(mesh);
+            mesh = GeometryFilters.reorderForPreVertexCache(mesh);
         }).toThrow();
     });
 
@@ -195,7 +195,7 @@ defineSuite([
             componentsPerAttribute : 3,
             values : [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17]
         };
-        MeshFilters.reorderForPreVertexCache(mesh);
+        GeometryFilters.reorderForPreVertexCache(mesh);
 
         expect(mesh.indexLists[0].values[0]).toEqual(0);
         expect(mesh.indexLists[0].values[1]).toEqual(1);
@@ -252,7 +252,7 @@ defineSuite([
                                                 maximumIndex : maximumIndex,
                                                 cacheSize : 24});
         expect(ACMRbefore).toBeGreaterThan(1.00);
-        mesh = MeshFilters.reorderForPostVertexCache(mesh);
+        mesh = GeometryFilters.reorderForPostVertexCache(mesh);
         indices = mesh.indexLists[0].values;
         var ACMRafter = Tipsify.calculateACMR({indices : indices,
                                                maximumIndex : maximumIndex,
@@ -280,7 +280,7 @@ defineSuite([
             }]
         };
 
-        var meshes = MeshFilters.fitToUnsignedShortIndices(mesh);
+        var meshes = GeometryFilters.fitToUnsignedShortIndices(mesh);
 
         expect(meshes.length).toEqual(1);
         expect(meshes[0]).toBe(mesh);
@@ -307,7 +307,7 @@ defineSuite([
             }]
         };
 
-        var meshes = MeshFilters.fitToUnsignedShortIndices(mesh);
+        var meshes = GeometryFilters.fitToUnsignedShortIndices(mesh);
 
         expect(meshes.length).toEqual(1);
         expect(meshes[0].attributes.time.componentDatatype).toEqual(ComponentDatatype.FLOAT);
@@ -346,7 +346,7 @@ defineSuite([
             }]
         };
 
-        var meshes = MeshFilters.fitToUnsignedShortIndices(mesh);
+        var meshes = GeometryFilters.fitToUnsignedShortIndices(mesh);
 
         expect(meshes.length).toEqual(2);
 
@@ -374,7 +374,7 @@ defineSuite([
         };
 
         expect(function() {
-            return MeshFilters.fitToUnsignedShortIndices(mesh);
+            return GeometryFilters.fitToUnsignedShortIndices(mesh);
         }).toThrow();
     });
 
@@ -401,7 +401,7 @@ defineSuite([
         };
 
         expect(function() {
-            return MeshFilters.fitToUnsignedShortIndices(mesh);
+            return GeometryFilters.fitToUnsignedShortIndices(mesh);
         }).toThrow();
     });
 
@@ -417,7 +417,7 @@ defineSuite([
             values : [p1.x, p1.y, p1.z, p2.x, p2.y, p2.z]
         };
 
-        mesh = MeshFilters.projectTo2D(mesh);
+        mesh = GeometryFilters.projectTo2D(mesh);
 
         var ellipsoid = Ellipsoid.WGS84;
         var projection = new GeographicProjection();
@@ -437,7 +437,7 @@ defineSuite([
         expect(mesh.attributes.position3D.values[5]).toEqual(p2.z);
     });
 
-    it('MeshFilters.encodeAttribute encodes positions', function() {
+    it('GeometryFilters.encodeAttribute encodes positions', function() {
         var c = new Cartesian3(-10000000.0, 0.0, 10000000.0);
         var encoded = EncodedCartesian3.fromCartesian(c);
 
@@ -450,7 +450,7 @@ defineSuite([
                 }
             }
         };
-        mesh = MeshFilters.encodeAttribute(mesh);
+        mesh = GeometryFilters.encodeAttribute(mesh);
 
         expect(mesh.attributes.positionHigh).toBeDefined();
         expect(mesh.attributes.positionHigh.values[0]).toEqual(encoded.high.x);
@@ -463,29 +463,29 @@ defineSuite([
         expect(mesh.attributes.position).not.toBeDefined();
     });
 
-    it('MeshFilters.encodeAttribute throws without a mesh', function() {
+    it('GeometryFilters.encodeAttribute throws without a mesh', function() {
         expect(function() {
-            MeshFilters.encodeAttribute(undefined);
+            GeometryFilters.encodeAttribute(undefined);
         }).toThrow();
     });
 
-    it('MeshFilters.encodeAttribute throws with mesh without attributes property', function() {
+    it('GeometryFilters.encodeAttribute throws with mesh without attributes property', function() {
         expect(function() {
-            MeshFilters.encodeAttribute({});
+            GeometryFilters.encodeAttribute({});
         }).toThrow();
     });
 
-    it('MeshFilters.encodeAttribute throws without attribute', function() {
+    it('GeometryFilters.encodeAttribute throws without attribute', function() {
         expect(function() {
             var mesh = {
                 attributes : {
                 }
             };
-            MeshFilters.encodeAttribute(mesh);
+            GeometryFilters.encodeAttribute(mesh);
         }).toThrow();
     });
 
-    it('MeshFilters.encodeAttribute throws without ComponentDatatype.FLOAT', function() {
+    it('GeometryFilters.encodeAttribute throws without ComponentDatatype.FLOAT', function() {
         expect(function() {
             var mesh = {
                 attributes : {
@@ -494,11 +494,11 @@ defineSuite([
                     values : [0.0]
                 }
             };
-            MeshFilters.encodeAttribute(mesh);
+            GeometryFilters.encodeAttribute(mesh);
         }).toThrow();
     });
 
-    it('MeshFilters.combine combines one mesh', function() {
+    it('GeometryFilters.combine combines one mesh', function() {
         var mesh = new Geometry({
             attributes : new GeometryAttribute({
                 position : {
@@ -509,11 +509,11 @@ defineSuite([
             })
         });
 
-        var combinedMesh = MeshFilters.combine([mesh]);
+        var combinedMesh = GeometryFilters.combine([mesh]);
         expect(combinedMesh).toBe(mesh);
     });
 
-    it('MeshFilters.combine combines several meshes', function() {
+    it('GeometryFilters.combine combines several meshes', function() {
         var mesh = new Geometry({
             attributes : {
                 position : new GeometryAttribute({
@@ -564,7 +564,7 @@ defineSuite([
             })]
         });
 
-        var combinedMesh = MeshFilters.combine([mesh, anotherMesh]);
+        var combinedMesh = GeometryFilters.combine([mesh, anotherMesh]);
         expect(combinedMesh).toEqual(new Geometry({
             attributes : {
                 position : new GeometryAttribute({
@@ -593,15 +593,15 @@ defineSuite([
         }));
     });
 
-    it('MeshFilters.combine throws with meshes', function() {
+    it('GeometryFilters.combine throws with meshes', function() {
         expect(function() {
-            MeshFilters.combine();
+            GeometryFilters.combine();
         }).toThrow();
     });
 
-    it('MeshFilters.combine throws when meshes.length is zero', function() {
+    it('GeometryFilters.combine throws when meshes.length is zero', function() {
         expect(function() {
-            MeshFilters.combine([]);
+            GeometryFilters.combine([]);
         }).toThrow();
     });